Rename getSeasonWeek to getChapterWeek and update related imports
