+++ src/features/game/events/landExpansion/deliver.ts
@@ -15,7 +15,7 @@ import {
   getSeasonalTicket,
 } from "features/game/types/chapters";
 import { NPCName } from "lib/npcs";
-import { getBumpkinHoliday } from "lib/utils/getSeasonWeek";
+import { getBumpkinHoliday } from "lib/utils/getChapterWeek";
 import { isWearableActive } from "features/game/lib/wearables";
 import { FACTION_OUTFITS } from "features/game/lib/factions";
 import { PATCH_FRUIT, PatchFruitName } from "features/game/types/fruits";
