+++ src/features/island/hud/Hud.tsx
@@ -22,7 +22,7 @@ import Decimal from "decimal.js-light";
 import { BuyCurrenciesModal } from "./components/BuyCurrenciesModal";
 import { MachineState } from "features/game/lib/gameMachine";
 import { useSound } from "lib/utils/hooks/useSound";
-import { SeasonBannerCountdown } from "./SeasonBannerCountdown";
+import { SeasonBannerCountdown } from "./ChapterBannerCountdown";
 import { TransactionCountdown } from "./Transaction";
 import { MarketplaceButton } from "./components/MarketplaceButton";
 import { PowerSkillsButton } from "./components/PowerSkillsButton";
