+++ src/features/world/ui/megastore/seasonalstore_components/ItemDetail.tsx
@@ -30,7 +30,7 @@ import {
   SeasonalStoreWearable,
   SeasonalTierItemName,
 } from "features/game/types/megastore";
-import { getItemDescription } from "../SeasonalStore";
+import { getItemDescription } from "../ChapterStore";
 import { getKeys } from "features/game/types/craftables";
 import { ARTEFACT_SHOP_KEYS } from "features/game/types/collectibles";
 import { SFLDiscount } from "features/game/lib/SFLDiscount";
